+++ src/main/java/gravit/code/badge/domain/Badge.java
@@ -1,7 +1,7 @@
 package gravit.code.badge.domain;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import gravit.code.badge.converter.JsonNodeConverter;
+import gravit.code.badge.support.JsonNodeConverter;
 import gravit.code.global.entity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
